requirements.txt
streamlit
opencv-python
mediapipe
deepface
numpy
pandas
matplotlib
Pillow
tensorflow
tf-keras
